requirements.txt
fastapi
redis-om
python-dotenv
sqlalchemy
psycopg2-binary
uvicorn
pika
requests
httpx